# Tuned H100/H200 triton fp8 block configs for fused_qkv_a_proj