<!-- markdownlint-disable -->

## Purpose

```
# Before
Benchmarking DeepSeek-V3, N=2112 K=7168
TFLOP/s comparison (block_size=(128, 128)):
WARNING 08-29 12:07:57 [fp8_utils.py:581] Using default W8A8 Block FP8 kernel config. Performance might be sub-optimal! Config file not found at /home/mgoin/code/vllm/vllm/model_executor/layers/quantization/utils/configs/N=2112,K=7168,device_name=NVIDIA_H100_80GB_HBM3,dtype=fp8_w8a8,block_shape=[128,128].json
BF16 vs W8A8 Block FP8 GEMMs:
   batch_size  torch-bf16  w8a8-block-fp8-triton
0         1.0    3.105397               0.527570
1        16.0   50.479968               9.050292
2        17.0   50.584079               9.802354
3       512.0  558.994182             215.180092
4      4096.0  711.448864             305.138408
5      4097.0  718.146297             306.611269
6     16381.0  704.885361             357.098143
7     16384.0  714.777622             354.501542

# After
Benchmarking DeepSeek-V3, N=2112 K=7168
TFLOP/s comparison (block_size=(128, 128)):
INFO 08-29 13:13:37 [fp8_utils.py:572] Using configuration from /home/mgoin/code/vllm/vllm/model_executor/layers/quantization/utils/configs/N=2112,K=7168,device_name=NVIDIA_H100_80GB_HBM3,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
BF16 vs W8A8 Block FP8 GEMMs:
   batch_size  torch-bf16  w8a8-block-fp8-triton
0         1.0    3.061336               1.303974
1        16.0   50.072462              20.520697
2        17.0   50.045805              21.695363
3       512.0  561.797319             307.285148
4      4096.0  728.706658             392.939256
5      4097.0  726.971636             342.251778
6     16381.0  705.874231             403.989800
7     16384.0  707.612514             403.475538
```

## Test Plan

## Test Result

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>